Fix NullReferenceException in Orphan() codepath
I ran into a `NullReferenceException` in this code path when locally debugging a function app in Visual Studio (source-linking brought me here). I don't know exactly what the side-effects of this null-ref are, but it seems wrong, especially since other parts of this method are correctly checking for a null `instance` method.